requirements.txt
aiogram==2.14
psycopg2==2.9.9
python-dotenv==1.0.1
